LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-cascade-layer-reorder-expected.txt


PASS When in the same layer, the last rule of each name wins
FAIL When in different layers, the rule of each name in the highest layer wins assert_equals: expected 100 but got 115
